Start sketching out RecipeBuilder datatypes + editors
We're gonna need to move very rapidly w/ RecipeBuilder to make the demo video work, so sketching out the way it handles and edits data ASAP seems wise. This is just a bunch of empty .cs files, but they're the empty .cs files we need for our data flow and the editors associated with it.